reorganized some type interfaces and cleaned up init.ts
